Move UnprivilegedToolResponse into api
